Routes getAllTeams and getAllSquad updated, still working - Matias Montes
