+++ api/src/routes/worldCupTeams/worldCupTeams.js
@@ -24,9 +24,12 @@ const getCoachDataApi = async function (teamId) {
     }
     })
 
+    if (!coachDataApiAux1.data.response[0].name) {
+      coachDataApiAux2 = 'There is no coach'
+    }else {
     coachDataApiAux2 = await coachDataApiAux1.data.response[0].name // the endpoint gets many coachs, so the main coach is the one into the array in position 0
 
-    if (coachDataApiAux2 === undefined) return 'There is no coach'
+    } 
    
     return coachDataApiAux2
     }
@@ -47,27 +50,29 @@ const getTeamsDataApi = async function () {
 
     /* console.log(AllTeamsDataApiAux1) */
 
-    AllTeamsDataApiAux2 = await AllTeamsDataApiAux1.map(el => {
-        let auxcoach = getCoachDataApi(el.team.id)
+    let auxArray = [];
 
+    AllTeamsDataApiAux2 = await AllTeamsDataApiAux1.map(el => {
+       /*  let auxcoach = await getCoachDataApi(el.team.id) */  //falta agregar coach
+        auxArray.push(el.team.id)
         return {
             id: el.team.id,
             name: el.team.name,
             logo: el.team.logo,
-            coach: auxcoach,
+            /* coach: auxcoach, */
             code: el.team.code,
             founded: el.team.founded
         };
     });
 
-    /* for (let i = 0; i < AllTeamsDataApiAux2.length; i++) {
-        AllTeamsDataApiAux2[i].coach = await getCoachDataApi(AllTeamsDataApiAux2[i].id)
-    } */
+     /*  for (let i = 0; i < auxArray.length; i++) {
+        AllTeamsDataApiAux2[i].coach = await getCoachDataApi(auxArray[i])
+    }   */
 
     return AllTeamsDataApiAux2
     }
 
-    router.get('/', async (req, res, next) => {
+     router.get('/allTeams', async (req, res, next) => {
 
         //idGroup is missing, it must be created in data base. Later search for it in DB
 
@@ -80,29 +85,65 @@ const getTeamsDataApi = async function () {
          catch (error) {
           next(error)
         } 
-      });
+      }); 
 
 
+      const getSquadDataApi = async function (id) {
 
-
-
-
-   /* router.get('/', async (req, res, next) => { */
-
-    /* console.log('hola') */
-     /* var API_KEYB = "993943c6a7d4a29527f6e9c92b7d0541"; */
-    /* try {
-      let A
-      A = await axios.get('https://v3.football.api-sports.io/teams?league=1&season=2022', {
-        headers: {          
-          "x-rapidapi-host": "v3.football.api-sports.io",
-          "x-rapidapi-key": `${API_KEY} `  
+        let AllSquadDataApiAux2
+        let AllSquadDataApiAux3
+    
+        let AllSquadDataApiAux1 = await axios.get(`https://v3.football.api-sports.io/players/squads?team=${id}`, {
+          headers: {          
+            "x-rapidapi-host": "v3.football.api-sports.io",
+            "x-rapidapi-key": `${API_KEY} `  
         }
-      })
-      res.status(200).send(A.data)
-    } catch (error) {
-      next(error)
-    }  */
-/*   }); */
+        })
+    
+        AllSquadDataApiAux1 = AllSquadDataApiAux1.data.response[0]
+        AllSquadDataApiAux3 = AllSquadDataApiAux1.team.id
+    
+         /* console.log(AllTeamsDataApiAux1)  */
+    
+        AllSquadDataApiAux2 = await AllSquadDataApiAux1.players.map(el => {
+            
+            return {
+                id: el.id,
+                name: el.name,
+                age: el.age,
+                /* number: el.number, */   // el campo suele estar en null por eso no lo envio
+                position: el.position,
+                photo: el.photo,
+                // el campo estado no va en la base de datos, se va a sacar
+                 id_team: AllSquadDataApiAux3
+            };
+        });
+    
+        return AllSquadDataApiAux2
+        }
+    
+        router.get('/squad/:id', async (req, res, next) => {
+    
+          /* console.log('hola') */
+    
+             const id = req.params.id;
+
+             /* console.log(id) */
+            //idGroup is missing, it must be created in data base. Later search for it in DB
+            try {
+              let A =  await getSquadDataApi(id);
+              console.log(A)  
+              
+              res.status(200).send(A)
+            }
+             catch (error) {
+              next(error)
+            }  
+          });
+
+
+
+
+ 
 
 module.exports = router;
